requirements/base.txt
django
psycopg2>=2.7,<3.0
djangorestframework
django-crontab
coreapi